pyproject.toml
[tool.poetry]
name = "lightning"
version = "0.1.0"
description = ""
authors = ["Jake Ledoux <contactjakeledoux@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
simpleeval = "^0.9.13"
pyglet = "^2.0.14"
hidapi = "^0.14.0"
pyserial = "^3.5"
dataclasses-json = "^0.6.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
